third commit

+++ src/br/com/dio/calculadora/Calculadora.java
@@ -15,7 +15,7 @@ public class Calculadora {
         int somar = somar(a, b);
         int subtrair = subtrair(a, b);
         int multiplicar = multiplicar(a, b);
-        int dividir = dividir(a, b);
+        double dividir = dividir(a, b);
 
         System.out.println("somar" + somar);
         System.out.println("subtrair" + subtrair);
@@ -35,7 +35,7 @@ public class Calculadora {
         return a*b;
     }
 
-    public static int dividir(int a, int b) {
+    public static double dividir(double a, double b) {
         return a/b;
     }
